Documents Prefix bug fixed Time format integer precision to float fixed System Prefix query join converted to left join

+++ src/System/App.php
@@ -260,7 +260,7 @@ class App
 		return (string) $number;
 	}
 
-	public function formatTime(int $time, ?bool $include_seconds = true): string
+	public function formatTime(float $time, ?bool $include_seconds = true): string
 	{
 		$neg    = $time < 0;
 		$time   = abs($time);
